+++ bcs-app/backend/apps/cluster/constants.py
@@ -41,28 +41,28 @@ PROJECT_ALL_CLUSTER = 'all'
 
 # 主机key的映射，便于前端进行展示
 CCHostKeyMappings = {
-    'bak_operator': 'BakOperator',
-    'classify_level_name': 'ClassifyLevelName',
-    'device_class': 'DeviceClass',
-    'device_type_id': 'DeviceType_Id',
-    'device_type_name': 'DeviceTypeName',
-    'hard_memo': 'HardMemo',
-    'host_id': 'HostID',
-    'host_name': 'HostName',
-    'idc': 'IDC',
-    'idc_area': 'idcArea',
-    'idc_area_id': 'idcAreaId',
-    'idc_id': 'IDC_ID',
-    'idcunit': 'IDCUnit',
-    'idcunit_id': 'IDCUnitID',
-    'inner_ip': 'InnerIP',
-    'memo': 'Memo',
-    'module_name': 'ModuleName',
-    'operator': 'Operator',
-    'osname': 'OSName',
-    'osversion': 'OSVersion',
-    'outer_ip': 'OuterIP',
-    'server_rack': 'serverRack',
+    'bak_operator': 'bk_bak_operator',
+    'classify_level_name': 'classify_level_name',
+    'device_class': 'svr_device_class',
+    'device_type_id': 'bk_svr_type_id',
+    'device_type_name': 'svr_type_name',
+    'hard_memo': 'hard_memo',
+    'host_id': 'bk_host_id',
+    'host_name': 'bk_host_name',
+    'idc': 'idc_name',
+    'idc_area': 'bk_idc_area',
+    'idc_area_id': 'bk_idc_area_id',
+    'idc_id': 'idc_id',
+    'idcunit': 'idc_unit_name',
+    'idcunit_id': 'idc_unit_id',
+    'inner_ip': 'bk_host_innerip',
+    'memo': 'bk_comment',
+    'module_name': 'module_name',
+    'operator': 'operator',
+    'osname': 'bk_os_name',
+    'osversion': 'bk_os_version',
+    'outer_ip': 'bk_host_outerip',
+    'server_rack': 'rack',
     'project_name': 'project_name',
     'cluster_name': 'cluster_name',
     'cluster_id': 'cluster_id',
